Remove require loops
