+++ flow/actions/performApiDataRequest.js
@@ -1,6 +1,6 @@
 // @flow
 
-import type { EndpointParams } from '../index';
+import type { EndpointParams } from '../types';
 import type { ActionCreator } from 'redux';
 import type { Action, ApiDataState } from '../reducer';
 import type { RequestHandler } from '../request';
