+++ flow/selectors/getResultData.js
@@ -1,6 +1,6 @@
 // @flow
 
-import type { EndpointParams } from '../index';
+import type { EndpointParams } from '../types';
 import { type ApiDataState } from '../reducer';
 
 declare export var getResultData: (
