+++ src/actions/invalidateApiDataRequest.ts
@@ -1,4 +1,4 @@
-import { EndpointParams } from '../index';
+import { EndpointParams } from '../types';
 import { getRequestKey } from '../helpers/getRequestKey';
 
 export interface InvalidateApiDataRequestAction {
