+++ src/helpers/createApiDataBinding.ts
@@ -1,8 +1,10 @@
-import { ApiDataState, EndpointParams, performApiRequest, getResultData, getApiDataRequest, ApiDataBinding, ApiDataRequest, invalidateApiDataRequest } from '..';
-import { Action } from 'redux';
-import { ThunkDispatch } from 'redux-thunk';
+import { EndpointParams, ApiDataBinding, ApiDataRequest } from '../types';
 import createApiDataRequest from './createApiDataRequest';
 import { getRequestKey } from './getRequestKey';
+import { ApiDataState } from '../reducer';
+import { getApiDataRequest } from '../selectors/getApiDataRequest';
+import { getResultData } from '../selectors/getResultData';
+import { Actions } from '../types';
 
 type BindingInstances = {
     [requestKey in string]: (apiData: ApiDataState, newApiDataRequest?: ApiDataRequest) => ApiDataBinding<any>;
@@ -10,10 +12,14 @@ type BindingInstances = {
 
 export class BindingsStore {
     bindingInstances: BindingInstances = {} as BindingInstances;
+    actions: Actions;
+
+    constructor(actions: Actions) {
+        this.actions = actions;
+    }
 
     getBinding(endpointKey: string, 
                params: EndpointParams = {}, 
-               dispatch: ThunkDispatch<{ apiData: ApiDataState }, void, Action>, 
                instanceId: string = '',
                apiData: ApiDataState, 
                request?: ApiDataRequest
@@ -22,8 +28,7 @@ export class BindingsStore {
         if (!this.bindingInstances[requestKey]) {
             this.bindingInstances[requestKey] = createApiDataBinding(
                 endpointKey, 
-                params, 
-                dispatch,
+                params,
                 this,
                 instanceId,
             );
@@ -32,10 +37,9 @@ export class BindingsStore {
     }
 }
 
-export const createApiDataBinding = (
+const createApiDataBinding = (
     endpointKey: string, 
     bindingParams: EndpointParams = {}, 
-    dispatch: ThunkDispatch<{ apiData: ApiDataState; }, void, Action>,
     bindingsStore: BindingsStore,
     instanceId: string = '',
 ): ((apiData: ApiDataState, request?: ApiDataRequest) => ApiDataBinding<any>) => {
@@ -46,9 +50,9 @@ export const createApiDataBinding = (
         request: request || getApiDataRequest(apiData, endpointKey, params, instanceId) || createApiDataRequest(endpointKey),
         perform: (performParams?: EndpointParams, body?: any) => {
             params = { ...bindingParams, ...performParams };
-            return dispatch(performApiRequest(endpointKey, params, body, instanceId, bindingsStore));
+            return bindingsStore.actions.perform(endpointKey, params, body, instanceId, bindingsStore);
         },
-        invalidateCache: () => dispatch(invalidateApiDataRequest(endpointKey, params, instanceId)),
-        getInstance: (newInstanceId: string) => bindingsStore.getBinding(endpointKey, params, dispatch, newInstanceId, apiData),
+        invalidateCache: () => bindingsStore.actions.invalidateCache(endpointKey, params, instanceId),
+        getInstance: (newInstanceId: string) => bindingsStore.getBinding(endpointKey, params, newInstanceId, apiData),
     });
 };
