+++ src/helpers/getRequestKey.ts
@@ -1,4 +1,4 @@
-import { EndpointParams } from '../index';
+import { EndpointParams } from '../types';
 
 export const getRequestKey = (endpointKey: string, params: EndpointParams = {}, instanceId: string = ''): string => 
     `${endpointKey}/${Object.keys(params).sort().map(param => `${param}=${params[param]}`).join('&')}${instanceId !== '' ? '#' + instanceId : ''}`;
