+++ src/selectors/cacheExpired.ts
@@ -1,4 +1,4 @@
-import { ApiDataEndpointConfig, ApiDataRequest } from '../index';
+import { ApiDataEndpointConfig, ApiDataRequest } from '../types';
 
 export const cacheExpired = (endpointConfig: ApiDataEndpointConfig, request: ApiDataRequest): boolean =>
     Date.now() - request.lastCall > (typeof endpointConfig.cacheDuration === 'number' ? endpointConfig.cacheDuration : Number.POSITIVE_INFINITY);
