+++ src/selectors/getResultData.ts
@@ -1,6 +1,6 @@
 import { denormalize } from 'normalizr';
 import { ApiDataState } from '../reducer';
-import { EndpointParams } from '../index';
+import { EndpointParams } from '../types';
 import { getApiDataRequest } from './getApiDataRequest';
 
 /**
