+++ src/withApiData.tsx
@@ -1,14 +1,15 @@
 import React from 'react';
-import { ApiDataBinding, ApiDataRequest, EndpointParams } from './index';
+import { ApiDataBinding, ApiDataRequest, EndpointParams, Actions } from './types';
 import { connect, ConnectedComponent } from 'react-redux';
 import { ApiDataState } from './reducer';
-import { getApiDataRequest, performApiRequest } from './index';
 import { Action } from './reducer';
 import hoistNonReactStatic from 'hoist-non-react-statics';
 import shallowEqual from 'shallowequal';
 import { ThunkDispatch } from 'redux-thunk';
 import { BindingsStore } from './helpers/createApiDataBinding';
-import { getActions, Actions } from './helpers/getActions';
+import { getActions } from './actions/performApiDataRequest';
+import { getApiDataRequest } from './selectors/getApiDataRequest';
+import { performApiRequest } from './actions/performApiDataRequest';
 
 type GetParams<TPropName extends string> = (ownProps: any, state: any) => { [paramName in TPropName]?: EndpointParams | EndpointParams[] };
 
@@ -118,10 +119,10 @@ export default function withApiData<TChildProps extends WithApiDataChildProps<TP
                 // check if we already have an instance of this bindingStore
                 let propNameBindingsStore: BindingsStore = this.bindingPropNameBindingsStore[propName];
                 if (propNameBindingsStore === undefined) {
-                    propNameBindingsStore = new BindingsStore();
+                    propNameBindingsStore = new BindingsStore(getActions(dispatch));
                     this.bindingPropNameBindingsStore[propName] = propNameBindingsStore;
                 }
-                return propNameBindingsStore.getBinding(endpointKey, params, dispatch, instanceId, apiData);
+                return propNameBindingsStore.getBinding(endpointKey, params, instanceId, apiData);
             }
 
             fetchDataIfNeeded() {
